tests refactoring + add test for DiscoverEnumerateAsync
